# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Task Management System API
  description: |
    REST API for the Serverless Task Management System.
    
    ## Authentication
    All endpoints except `/health` require a valid JWT token in the Authorization header.
    Tokens are obtained from AWS Cognito after successful authentication.
    
    ## Roles
    - **Admin**: Full CRUD access to all resources
    - **Member**: Read access to assigned tasks, can update task status only
    
    ## Email Domain Restriction
    Only users with `@amalitech.com` or `@amalitechtraining.org` emails can register.
  version: 1.0.0
  contact:
    name: Prince Ayiku
    email: prince@amalitech.com
servers:
  - url: https://api.taskmanager.amalitech.com/api/v1
    description: Production
  - url: https://api-staging.taskmanager.amalitech.com/api/v1
    description: Staging
  - url: https://api-dev.taskmanager.amalitech.com/api/v1
    description: Development
...
paths:
...
  /tasks/{taskId}:
    parameters:
      - name: taskId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [Tasks]
      summary: Get task by ID
      description: Returns a single task by ID
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Tasks]
      summary: Update task
      description: Admin only. Updates all task fields.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Tasks]
      summary: Delete task
      description: Admin only. Deletes a task.
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Task deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from AWS Cognito
  schemas:
...
    TaskStatus:
      type: string
      enum: [OPEN, IN_PROGRESS, UNDER_REVIEW, COMPLETED, CLOSED]
    Priority:
      type: string
      enum: [LOW, MEDIUM, HIGH, URGENT]
...
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        priority:
          $ref: '#/components/schemas/Priority'
        status:
          $ref: '#/components/schemas/TaskStatus'
        dueDate:
          type: string
          format: date-time
...
    UserSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        email:
          type: string
          format: email
...
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/TaskStatus'
        priority:
          $ref: '#/components/schemas/Priority'
        createdBy:
          $ref: '#/components/schemas/UserSummary'
        assignedTo:
          type: array
          items:
            $ref: '#/components/schemas/UserSummary'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        dueDate:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
        timestamp:
          type: string
          format: date-time
        path:
          type: string
  responses:
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: BadRequest
            message: Task title is required
            details:
              field: title
              constraint: required
            timestamp: '2026-02-06T10:00:00Z'
            path: /api/v1/tasks
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            message: Missing or invalid authentication token
            timestamp: '2026-02-06T10:00:00Z'
            path: /api/v1/tasks
    Forbidden:
      description: Valid token but insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Forbidden
            message: You do not have permission to perform this action
            timestamp: '2026-02-06T10:00:00Z'
            path: /api/v1/tasks
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: NotFound
            message: Task not found
            timestamp: '2026-02-06T10:00:00Z'
            path: /api/v1/tasks/123
...